+++ extensions/google-drive/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Google Drive Changelog
 
+## [Faster indexing with progress] - 2022-05-06
+
+- Show the indexing progress in percentage + processed/total format.
+
 ## [Toggle details] - 2022-05-03
 
 - Added a general action to toggle the right-side details view.